Move LatencyMeasurer
